requirements.txt
astroid==1.4.5
colorama==0.3.7
coverage==4.0.3
Flask==0.10.1
Flask-Excel==0.0.4
Flask-MySQLdb==0.2.0
Flask-Testing==0.4.2
Flask-Uploads==0.2.0
Flask-WTF==0.12
itsdangerous==0.24
Jinja2==2.8
lazy-object-proxy==1.2.2
logilab-common==1.2.0
MarkupSafe==0.23
mysqlclient==1.3.7
nose==1.3.7
nosexcover==1.0.10
pyexcel==0.2.0
pyexcel-io==0.1.0
pyexcel-webio==0.0.6
pylint==1.5.5
pylint-flask==0.3
pylint-plugin-utils==0.2.3
simplejson==3.8.2
six==1.10.0
texttable==0.8.4
Werkzeug==0.11.4
wrapt==1.10.8
WTForms==2.1
